requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
asyncio==3.4.3
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pycodestyle==2.10.0
python-dotenv==0.21.0
python-telegram-bot==20.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
